backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
apscheduler==3.10.4
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
pydantic==2.5.2
pydantic[email]==2.5.2
